Rename actoinDelegator.js
